Fix scheduling.

+++ state.sql
@@ -69,7 +69,7 @@ CREATE TABLE IF NOT EXISTS queue (
    suite suite not null,
    command text not null,
    committer text null,
-   priority integer default 0 not null,
+   priority bigint default 0 not null,
    foreign key (package) references package(name),
    -- Some subworker-specific indication of what we are expecting to do.
    context text,
